LayoutTests/imported/w3c/web-platform-tests/html/semantics/embedded-content/the-img-element/currentSrc-blob-cache-expected.txt

 

PASS currentSrc is right even if underlying image is a shared blob 
